+++ script/files.lua
@@ -7,6 +7,8 @@ local proto    = require 'proto'
 local lang     = require 'language'
 local await    = require 'await'
 local timer    = require 'timer'
+local plugin   = require 'plugin'
+local util     = require 'utility'
 
 local m = {}
 
@@ -22,13 +24,23 @@ m.globalVersion = 0
 m.linesMap = setmetatable({}, { __mode = 'v' })
 m.astMap   = setmetatable({}, { __mode = 'v' })
 
+local uriMap = {}
+local function getUriKey(uri)
+    if not uriMap[uri] then
+        if platform.OS == 'Windows' then
+            uriMap[uri] = uri:lower()
+        else
+            uriMap[uri] = uri
+        end
+    end
+    return uriMap[uri]
+end
+
 --- 打开文件
 ---@param uri uri
 function m.open(uri)
     local originUri = uri
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     m.openMap[uri] = true
     m.onWatch('open', originUri)
 end
@@ -37,9 +49,7 @@ end
 ---@param uri uri
 function m.close(uri)
     local originUri = uri
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     m.openMap[uri] = nil
     m.onWatch('close', originUri)
 end
@@ -48,33 +58,25 @@ end
 ---@param uri uri
 ---@return boolean
 function m.isOpen(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     return m.openMap[uri] == true
 end
 
 --- 标记为库文件
 function m.setLibraryPath(uri, libraryPath)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     m.libraryMap[uri] = libraryPath
 end
 
 --- 是否是库文件
 function m.isLibrary(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     return m.libraryMap[uri] ~= nil
 end
 
 --- 获取库文件的根目录
 function m.getLibraryPath(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     return m.libraryMap[uri]
 end
 
@@ -85,16 +87,12 @@ end
 --- 是否存在
 ---@return boolean
 function m.exists(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     return m.fileMap[uri] ~= nil
 end
 
 function m.asKey(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     return uri
 end
 
@@ -105,10 +103,9 @@ function m.setText(uri, text)
     if not text then
         return
     end
+    --log.debug('setText', uri)
     local originUri = uri
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local create
     if not m.fileMap[uri] then
         m.fileMap[uri] = {
@@ -116,6 +113,11 @@ function m.setText(uri, text)
             version = 0,
         }
         create = true
+        m._pairsCache = nil
+    end
+    local suc, newText = plugin.dispatch('OnSetText', originUri, text)
+    if suc then
+        text = newText
     end
     local file = m.fileMap[uri]
     if file.text == text then
@@ -137,9 +139,7 @@ end
 
 --- 获取文件版本
 function m.getVersion(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri]
     if not file then
         return nil
@@ -151,9 +151,7 @@ end
 ---@param uri uri
 ---@return string text
 function m.getText(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri]
     if not file then
         return nil
@@ -165,14 +163,13 @@ end
 ---@param uri uri
 function m.remove(uri)
     local originUri = uri
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri]
     if not file then
         return
     end
     m.fileMap[uri] = nil
+    m._pairsCache = nil
 
     m.globalVersion = m.globalVersion + 1
     await.close('files.version')
@@ -183,6 +180,7 @@ end
 function m.removeAll()
     m.globalVersion = m.globalVersion + 1
     await.close('files.version')
+    m._pairsCache = nil
     for uri in pairs(m.fileMap) do
         if not m.libraryMap[uri] then
             m.fileMap[uri]  = nil
@@ -198,6 +196,7 @@ end
 function m.removeAllClosed()
     m.globalVersion = m.globalVersion + 1
     await.close('files.version')
+    m._pairsCache = nil
     for uri in pairs(m.fileMap) do
         if  not m.openMap[uri]
         and not m.libraryMap[uri] then
@@ -210,13 +209,30 @@ function m.removeAllClosed()
     --m.notifyCache = {}
 end
 
+--- 获取一个包含所有文件uri的数组
+---@return uri[]
+function m.getAllUris()
+    local files = m._pairsCache
+    local i = 0
+    if not files then
+        files = {}
+        m._pairsCache = files
+        for uri in pairs(m.fileMap) do
+            i = i + 1
+            files[i] = uri
+        end
+    end
+    return m._pairsCache
+end
+
 --- 遍历文件
 function m.eachFile()
-    local map = {}
-    for uri, file in pairs(m.fileMap) do
-        map[uri] = file
+    local files = m.getAllUris()
+    local i = 0
+    return function ()
+        i = i + 1
+        return files[i]
     end
-    return pairs(map)
 end
 
 --- Pairs dll files
@@ -269,7 +285,12 @@ function m.compileAst(uri, text)
     if state then
         state.uri = uri
         state.ast.uri = uri
+        local clock = os.clock()
         parser:luadoc(state)
+        local passed = os.clock() - clock
+        if passed > 0.1 then
+            log.warn(('Parse LuaDoc of [%s] takes [%.3f] sec, size [%.3f] kb.'):format(uri, passed, #text / 1000))
+        end
         return state
     else
         log.error(err)
@@ -281,9 +302,7 @@ end
 ---@param uri uri
 ---@return table ast
 function m.getAst(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     if uri ~= '' and not m.isLua(uri) then
         return nil
     end
@@ -304,9 +323,7 @@ end
 ---@param uri uri
 ---@return table lines
 function m.getLines(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri]
     if not file then
         return nil
@@ -321,9 +338,7 @@ end
 
 --- 获取原始uri
 function m.getOriginUri(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri] or m.dllMap[uri]
     if not file then
         return nil
@@ -332,29 +347,25 @@ function m.getOriginUri(uri)
 end
 
 function m.getUri(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     return uri
 end
 
 --- 获取文件的自定义缓存信息（在文件内容更新后自动失效）
 function m.getCache(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri]
     if not file then
         return nil
     end
-    file.cacheActiveTime = timer.clock()
+    --file.cacheActiveTime = timer.clock()
     return file.cache
 end
 
 --- 判断文件名相等
 function m.eq(a, b)
     if platform.OS == 'Windows' then
-        return a:lower() == b:lower()
+        return a:lower():gsub('[/\\]+', '/') == b:lower():gsub('[/\\]+', '/')
     else
         return a == b
     end
@@ -422,10 +433,7 @@ function m.saveDll(uri, content)
     if not content then
         return
     end
-    local luri = uri
-    if platform.OS == 'Windows' then
-        luri = uri:lower()
-    end
+    local luri = getUriKey(uri)
     local file = {
         uri   = uri,
         opens = {},
@@ -454,9 +462,7 @@ end
 ---@param uri uri
 ---@return string[]|nil
 function m.getDllOpens(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.dllMap[uri]
     if not file then
         return nil
@@ -468,9 +474,7 @@ end
 ---@param uri uri
 ---@return string[]|nil
 function m.getDllWords(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.dllMap[uri]
     if not file then
         return nil
@@ -499,9 +503,7 @@ function m.flushCache()
 end
 
 function m.flushFileCache(uri)
-    if platform.OS == 'Windows' then
-        uri = uri:lower()
-    end
+    uri = getUriKey(uri)
     local file = m.fileMap[uri]
     if not file then
         return